refactor(Product): repository 폴더 생성
