+++ src/main/java/gift/domain/product/ProductService.java
@@ -1,8 +1,8 @@
 package gift.domain.product;
 
+import gift.domain.product.repository.JpaProductRepository;
 import gift.global.exception.BusinessException;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.jdbc.core.JdbcTemplate;
